ui/change

+++ storage/framework/views/a6c20f767fef42ca9a5cd402fae47651.php
@@ -1,19 +1,19 @@
 <div class="flex items-center">
     <div class="ml-4 pointer-events-auto">
         <button wire:click="toggleLike" class="focus:outline-none">
-            <!--[if BLOCK]><![endif]--><?php if($isLiked): ?>
+            <?php if($isLiked): ?>
                 <i class="fa-solid fa-heart text-red-500 <?php echo e($viewType === 'list' ? 'text-xl' : 'text-2xl'); ?> mr-2"></i>
             <?php else: ?>
                 <i class="fa-regular fa-heart 
                 <?php echo e($viewType === 'list' ? 'text-white text-xl' : 'text-gray-400 text-2xl'); ?> 
                 hover:text-red-500 mr-2"></i>
-            <?php endif; ?><!--[if ENDBLOCK]><![endif]-->
+            <?php endif; ?>
         </button>
     </div>
     
     
     <div class="mr-2">
-        <!--[if BLOCK]><![endif]--><?php if($viewType === 'list'): ?>
+        <?php if($viewType === 'list'): ?>
             <span class="text-white"><?php echo e($likeCount); ?></span>
         <?php elseif($viewType === 'show'): ?>
             <div x-data="{open:false}">
@@ -35,28 +35,30 @@
                         <hr class="border-green-500 border-1">
                         
                         <div class="mx-auto h-full mt-8">
-                            <!--[if BLOCK]><![endif]--><?php $__currentLoopData = $post->likes; $__env->addLoop($__currentLoopData); foreach($__currentLoopData as $like): $__env->incrementLoopIndices(); $loop = $__env->getLastLoop(); ?>
+                            <?php $__currentLoopData = $post->likes; $__env->addLoop($__currentLoopData); foreach($__currentLoopData as $like): $__env->incrementLoopIndices(); $loop = $__env->getLastLoop(); ?>
                                 <div class="flex items-center justify-between bg-white rounded-lg shadow p-4 mb-4 hover:bg-gray-50 transition">
                                     <a href="" class="flex items-center space-x-4 w-full ml-2">
-                                        <!--[if BLOCK]><![endif]--><?php if($like->user->avatar): ?>
+                                        <?php if(optional($like->user)->avatar): ?>
                                             <img src="<?php echo e($like->user->avatar); ?>" alt="<?php echo e($like->user->name); ?>" class="w-12 h-12 rounded-full object-cover">
                                         <?php else: ?>
-                                        <div class="w-12 h-12 bg-gray-300 rounded-full flex items-center justify-center text-white font-bold">
-                                            <?php echo e(strtoupper(substr($like->user->name, 0, 1))); ?>
+                                            <?php if(optional($like->user)->name): ?>
+                                                <div class="w-12 h-12 bg-gray-300 rounded-full flex items-center justify-center text-white font-bold">
+                                                    <?php echo e(strtoupper(substr($like->user->name, 0, 1))); ?>
 
-                                        </div>
-                                        <?php endif; ?><!--[if ENDBLOCK]><![endif]-->
+                                                </div>
+                                                <div class="flex flex-row sm:flex-row sm:items-center sm:justify-center text-center sm:space-x-2 ">
+                                                    <p class="font-semibold text-2xl truncate "><?php echo e($like->user->name); ?></p>
+                                                </div>
+                                            <?php endif; ?>
+                                        <?php endif; ?>
 
-                                        <div class="flex flex-row sm:flex-row sm:items-center sm:justify-center text-center sm:space-x-2 ">
-                                            <p class="font-semibold text-2xl truncate "><?php echo e($like->user->name); ?></p>
-                                        </div>
                                     </a>
                                 </div>
-                            <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?><!--[if ENDBLOCK]><![endif]-->
+                            <?php endforeach; $__env->popLoop(); $loop = $__env->getLastLoop(); ?>
                         </div>
                     </div>
                 </div>
             </div>
-        <?php endif; ?><!--[if ENDBLOCK]><![endif]-->
+        <?php endif; ?>
     </div>
 </div><?php /**PATH /Users/kawaimayu/Desktop/gi_12_tripe_may/resources/views/livewire/post-like.blade.php ENDPATH**/ ?>
